pyproject.toml
[project]
name = "agent-neo"
version = "0.1.0"
description = "Python agent using LiteLLM with local hailo-ollama LLM"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "strands-agents>=1.27.0",
]

[project.scripts]
agent-neo = "agent_neo.main:main"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agent_neo"]
